attachments/requirement.txt
streamlit
pandas
bcrypt
google-genai
fpdf
plotly
openpyxl